Removed compilation warnings (except linker warnings related to external libs). Switching some data types to match without implicitly narrowing or expanding them in operations. Resolved a case of a bad template code in ImG_debuger::addDebugValueToFolder method. Changed that template impl from macro injection to impl via .inl file.

+++ OpenGLTester/source/BaseClasses/Public/IndexBuffer.h
@@ -9,27 +9,27 @@ class IndexBuffer
 {
 public:
 
-    IndexBuffer(const unsigned int* _data, const unsigned int _count, GLenum _memUsageType = GL_STATIC_DRAW);
-    IndexBuffer(const unsigned int _count = 1024);
+    IndexBuffer(const unsigned int* _data, const size_t _count, GLenum _memUsageType = GL_STATIC_DRAW);
+    IndexBuffer(const size_t _count = 1024);
     ~IndexBuffer();
 
     //Primary method for adding indices to draw for batch-rendering method
     void addRenderObject(const SPTR<RenderObject>& _RenderObject);
     //Only use this one if you adjusted your indices elsewhere or it's the first batch of indices! Avoid if possible!
-    void push_back_drawIndices(std::vector<unsigned int> _indices);
+    void push_back_drawIndices(std::vector<size_t> _indices);
 
     void bind() const;
     void unbind() const;
     
-    unsigned int getCount() const { return m_count; }
+    size_t getCount() const { return m_count; }
 
     //TODO: add methods for changing data, for "locking" buffer for data streaming
     
 private:
     
     GLuint m_rendererId;
-    unsigned int m_count;
-    unsigned int m_adjustment = 0;
+    size_t m_count;
+    size_t m_adjustment = 0;
 
-    std::vector<unsigned int> m_vertexIndicesPool; 
+    std::vector<size_t> m_vertexIndicesPool; 
 };
